requirements.txt
Flask
google-cloud-bigquery
gunicorn